### 2022-02-04 #4132 @whedon
[PRE REVIEW]: LangToGroup: A Haskell package for group presentation construction from formal language
**Submitting author:** @MaxVortman (<a href="http://orcid.org/0000-0003-0730-494X">Maksym Shamrai</a>)
**Repository:** <a href="https://github.com/YaccConstructor/LangToGroup" target ="_blank">https://github.com/YaccConstructor/LangToGroup</a>
**Version:** v1.0.0
