### 2022-02-05 #4132 @Kevin-Mattheus-Moerman
@MaxVortman can you check those potentially missing/invalid DOI's :point_up: you can update the paper here by calling: `@whedon generate pdf`. And you can also check the references by calling `@whedon check references`.
