### 2022-02-06 #4132 @Nikoleta-v3
Thank you @MaxVortman! Could you:

- 1️⃣ check for potentially missing/invalid DOI's https://github.com/openjournals/joss-reviews/issues/4132#issuecomment-1030581850
- 2️⃣ take a look through https://bit.ly/joss-reviewers and suggest at least four potential reviewers 😄
