### 2022-02-11 #4132 @Nikoleta-v3
Thank you for the suggestions @MaxVortman 👍🏻

Regarding the DOI. My apologies for the unclear message what I meant was can you fix the potential missing/invalid DOIs of your references ⬆️
