+++ paper/paper.bib
@@ -2,12 +2,14 @@
   author       = "Chomsky, N.",
   title        = "Some methodological remarks on generative grammar",
   year         = 1961,
+  doi={10.1080/00437956.1961.11659755},
   publisher  = {Aspects of the Theory of Syntax}
 }
 @book{chomsky2,
   author       = "Chomsky, N.",
   title        = "Syntactic structures",
   year         = 2002,
+  doi={10.1515/9783110218329},
   publisher  = {Walter de Gruyter}
 }
 @inproceedings{Azimov,
@@ -26,9 +28,10 @@ keywords = {transitive closure, context-free path querying, context-free grammar
 location = {Houston, Texas},
 series = {GRADES-NDA ’18}
 }
-@inproceedings{Hellings2015QueryingFP,
+@article{Hellings2015QueryingFP,
   title={Querying for Paths in Graphs using Context-Free Path Queries},
   author={Jelle Hellings},
+  journal={arXiv: Databases},
   year={2015}
 }
 @article{Huber2007,
@@ -58,6 +61,7 @@ url={https://doi.org/10.1186/1471-2105-8-S6-S8}
   number={1},
   pages={109--127},
   year={1988},
+  doi={10.1177/0038038588022001007},
   publisher={British Sociological Association Publications Limited}
 }
 @article{ConjGrammars,
@@ -74,7 +78,8 @@ journal = {J. Autom. Lang. Comb.},
 month = apr,
 pages = {519–535},
 numpages = {17},
-keywords = {parsing, intersection, conjuctive grammar, descriptional complexity, context-free grammar}
+keywords = {parsing, intersection, conjuctive grammar, descriptional complexity, context-free grammar},
+doi={10.25596/jalc-2001-519}
 }
 @article{BoolGrammars,
 title = "Boolean grammars",
@@ -84,7 +89,7 @@ number = "1",
 pages = "19 - 48",
 year = "2004",
 issn = "0890-5401",
-doi = "https://doi.org/10.1016/j.ic.2004.03.006",
+doi = {10.1016/j.ic.2004.03.006},
 url = "http://www.sciencedirect.com/science/article/pii/S0890540104001075",
 author = "Alexander Okhotin",
 keywords = "Context-free grammar, Intersection, Complement, Language equation, Parsing, Conjunctive grammar, Trellis automaton, Cellular automaton",
@@ -100,7 +105,8 @@ abstract = "A new generalization of context-free grammars is introduced: Boolean
  publisher = {Annals of Mathematics},
  title = {Isoperimetric and Isodiametric Functions of Groups},
  volume = {156},
- year = {2002}
+ year = {2002},
+ doi={10.2307/3597195}
 }
 @incollection{rotman1995word,
   title={The Word Problem},
@@ -108,7 +114,8 @@ abstract = "A new generalization of context-free grammars is introduced: Boolean
   booktitle={An Introduction to the Theory of Groups},
   pages={418--470},
   year={1995},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-1-4612-4176-8}
 }
 @manual{GAP4,
   key          = "GAP",
