### 2022-02-12 #4132 @Nikoleta-v3
👋 @ritwikagarwal & @JoyMitra & @adk9 & @ocramz would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "LangToGroup: A Haskell package for group presentation construction from formal language". You can find more information at the top of this Github issue 🆙
