### 2022-02-20 #4132 @JoyMitra
Hi,

I will be happy to help.

Best,
Joy Mitra

On Sat, Feb 12, 2022, 5:35 AM Nikoleta Glynatsi ***@***.***>
wrote:

> 👋 @ritwikagarwal <https://github.com/ritwikagarwal> & @JoyMitra
> <https://github.com/JoyMitra> & @adk9 <https://github.com/adk9> & @ocramz
> <https://github.com/ocramz> would any of you be willing to review this
> submission for JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow
> these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
> The submission I'd like you to review is titled: "LangToGroup: A Haskell
> package for group presentation construction from formal language". You can
> find more information at the top of this Github issue 🆙
>
> Please let me know if you're available 😄 Thank you!
>
