### 2022-02-20 #4184 @whedon
[REVIEW]: LangToGroup: A Haskell package for group presentation construction from formal language
**Submitting author:** @MaxVortman (<a href="http://orcid.org/0000-0003-0730-494X">Maksym Shamrai</a>)
**Repository:** https://github.com/YaccConstructor/LangToGroup
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
