### 2022-02-26 YaccConstructor/LangToGroup#49 @ocramz
[JOSS] Contribution guidelines
It would be nice to state some contribution guidelines, e.g.

- Where should a newcomer to this project start ?

Tracking JOSS issue https://github.com/openjournals/joss-reviews/issues/4184
