### 2022-02-26 YaccConstructor/LangToGroup#50 @ocramz
[JOSS] Statement of need could be improved
I appreciate this is a fairly niche and advanced CS topic, but the authors could elaborate a bit more on where a connection between formal languages and group theory might be needed.

Tracking https://github.com/openjournals/joss-reviews/issues/4184
