### 2022-02-26 #4184 @ocramz
I congratulate the authors on a very interesting contribution. In my opinion the paper, in its current form, could elaborate a bit more on the statement of need/"applications" part; it connects two areas that are rich in theory, and I would guess there's plenty of applications that could benefit from this connection.

@Nikoleta-v3 I suggest to relax the review timeline; Kyiv is currently being invaded by Russia, and the first author lives there.
