### 2023-02-01 #4184 @Nikoleta-v3
Dear @ocramz and @JoyMitra. I would like to thank you again for agreeing to review this submission for JOSS 🙏

I just spoke to the author; they are doing okay and most importantly they are safe 😊  They have decided to withdraw the submission so I am going to go ahead and do this.

Once again thank you for all your time!
